+++ CHANGES.md
@@ -9,7 +9,7 @@
 
 ## Done
 
-- Fix: do not record init and destroy related actions
+- Fix: init and destroy are not called during hot-swaping
 - Fix generic propagation name argument
 - Fix lifecycle ordering
 - Add init and destroy lifecycle hooks